# tests/fixtures/legacy-openapi-7006aec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Audio Analysis API",
    "description": "High-quality music analysis using Essentia C++ core via Python.",
    "version": "4.1.0"
  },
  "paths": {
    "/analyze/rhythm": {
      "post": {
        "tags": [
          "Analysis"
        ],
        "summary": "Analyze Rhythm",
        "description": "Extract BPM, beats, confidence, and high-quality onsets.",
        "operationId": "analyze_rhythm_analyze_rhythm_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_rhythm_analyze_rhythm_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RhythmAnalysis"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/analyze/structure": {
      "post": {
        "tags": [
          "Analysis"
        ],
        "summary": "Analyze Structure",
        "description": "Segment audio into sections (intro, verse, chorus, etc.) using SBic.",
        "operationId": "analyze_structure_analyze_structure_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_structure_analyze_structure_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StructureAnalysis"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/analyze/structure/allin1": {
      "post": {
        "tags": [
          "Analysis"
        ],
        "summary": "Analyze Structure Allin1",
        "description": "Functional song structure (intro/verse/chorus/...) via mir-aidj/all-in-one.",
        "operationId": "analyze_structure_allin1_analyze_structure_allin1_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_structure_allin1_analyze_structure_allin1_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StructureAnalysis"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/analyze/classification": {
      "post": {
        "tags": [
          "Analysis"
        ],
        "summary": "Analyze Classification",
        "description": "Analyze Genre, Mood, Tags, and additional classification features using Essentia TensorFlow models.",
        "operationId": "analyze_classification_analyze_classification_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "features",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated list of features to analyze. Available: genre, mood, tags, danceability, approachability, engagement, acoustic_electronic, bright_dark, instrument, tonal_atonal. Default: all.",
              "title": "Features"
            },
            "description": "Comma-separated list of features to analyze. Available: genre, mood, tags, danceability, approachability, engagement, acoustic_electronic, bright_dark, instrument, tonal_atonal. Default: all."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_classification_analyze_classification_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClassificationAnalysis"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analyze/tonal": {
      "post": {
        "tags": [
          "Analysis"
        ],
        "summary": "Analyze Tonal",
        "description": "Extract key, scale, deep-learning tempo (TempoCNN), and pitch (CREPE).",
        "operationId": "analyze_tonal_analyze_tonal_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_tonal_analyze_tonal_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnhancedTonalAnalysis"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/analyze/tonal/key": {
      "post": {
        "tags": [
          "Analysis"
        ],
        "summary": "Analyze Tonal Key",
        "description": "Extract key, scale, and key strength only.",
        "operationId": "analyze_tonal_key_analyze_tonal_key_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_tonal_key_analyze_tonal_key_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TonalAnalysis"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/analyze/tonal/tempo": {
      "post": {
        "tags": [
          "Analysis"
        ],
        "summary": "Analyze Tonal Tempo",
        "description": "Extract deep-learning tempo only (TempoCNN).",
        "operationId": "analyze_tonal_tempo_analyze_tonal_tempo_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_tonal_tempo_analyze_tonal_tempo_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TempoAnalysis"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/analyze/tonal/pitch": {
      "post": {
        "tags": [
          "Analysis"
        ],
        "summary": "Analyze Tonal Pitch",
        "description": "Extract pitch only (CREPE).",
        "operationId": "analyze_tonal_pitch_analyze_tonal_pitch_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_tonal_pitch_analyze_tonal_pitch_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PitchAnalysis"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/analyze/vocals": {
      "post": {
        "tags": [
          "Analysis"
        ],
        "summary": "Analyze Vocals",
        "description": "Detect whether audio contains vocals or is instrumental.",
        "operationId": "analyze_vocals_analyze_vocals_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_vocals_analyze_vocals_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VocalAnalysis"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/analyze/fast": {
      "post": {
        "tags": [
          "Analysis"
        ],
        "summary": "Analyze Fast",
        "description": "Fast first-pass analysis for interactive editing.\n\nReturns only rhythm, energy, and song structure. This is enough to build\nbeat/section previews without waking classifier, tonal, pitch, or vocal\nTensorFlow models on the blocking upload path.",
        "operationId": "analyze_fast_analyze_fast_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_fast_analyze_fast_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FastAnalysis"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/analyze/full": {
      "post": {
        "tags": [
          "Analysis"
        ],
        "summary": "Analyze Full",
        "description": "Perform full rhythm, structural, classification, tonal, and vocal analysis.",
        "operationId": "analyze_full_analyze_full_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_full_analyze_full_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FullAnalysis"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_analyze_classification_analyze_classification_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_analyze_classification_analyze_classification_post"
      },
...
      "Body_analyze_rhythm_analyze_rhythm_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_analyze_rhythm_analyze_rhythm_post"
      },
      "Body_analyze_structure_allin1_analyze_structure_allin1_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_analyze_structure_allin1_analyze_structure_allin1_post"
      },
      "Body_analyze_structure_analyze_structure_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_analyze_structure_analyze_structure_post"
      },
      "Body_analyze_tonal_analyze_tonal_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_analyze_tonal_analyze_tonal_post"
      },
      "Body_analyze_tonal_key_analyze_tonal_key_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_analyze_tonal_key_analyze_tonal_key_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RhythmAnalysis": {
        "properties": {
          "bpm": {
            "type": "number",
            "title": "Bpm"
          },
          "beats": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Beats"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "onsets": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Onsets"
          },
          "duration": {
            "type": "number",
            "title": "Duration"
          },
          "energy": {
            "$ref": "#/components/schemas/EnergyData"
          }
        },
        "type": "object",
        "required": [
          "bpm",
          "beats",
          "confidence",
          "onsets",
          "duration",
          "energy"
        ],
        "title": "RhythmAnalysis"
      },
...
      "StructureAnalysis": {
        "properties": {
          "sections": {
            "items": {
              "$ref": "#/components/schemas/Section"
            },
            "type": "array",
            "title": "Sections"
          },
          "boundaries": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Boundaries"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "unknown"
          },
          "analyzed_duration_s": {
            "type": "number",
            "title": "Analyzed Duration S",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "sections",
          "boundaries"
        ],
        "title": "StructureAnalysis"
      },
...
      "TonalAnalysis": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "scale": {
            "type": "string",
            "title": "Scale"
          },
          "strength": {
            "type": "number",
            "title": "Strength"
          }
        },
        "type": "object",
        "required": [
          "key",
          "scale",
          "strength"
        ],
        "title": "TonalAnalysis"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "description": "API Key required for authentication. Contact the administrator to receive your API key.",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}